+++ src/entry.jsx
@@ -4,26 +4,26 @@ import { Provider } from 'react-redux';
 import { ReduxRouter } from 'redux-router';
 import configureStore from './store/configureStore';
 
+//Object.assign() polyfill for IE (used in the reducer)
+import './es6-polyfill.js';
+
 import getGithubProjects from './getInitialData';
 
+//STEP 1: get project data from the static JSON file hosted on the CDN
 getGithubProjects().then( json => startRedux(json) );
 
+//STEP 2: Launch the Redux application once we get data
 function startRedux(githubProjects) {
 
   const initialState = {
-    staticContent: {
-        projectName: 'bestof222.js.org',
-        repo: 'https://github.com/michaelrambeau/bestofjs-webui'
-    },
     githubProjects
   };
-  console.log('initialState', initialState);
 
   const store = configureStore(initialState);
 
-  if (true) render(
+  render(
     <Provider store={store}>
-      <ReduxRouter mike="xxx" />
+      <ReduxRouter />
     </Provider>,
     window.document.getElementById('app')
   );